clean up staging the object/resource db
